+++ src/auto_segmentation/src/main.py
@@ -2,28 +2,48 @@ import os
 from librosa.core import load
 import numpy as np
 import scipy as sc
-from DataParsing import*
-from SVM import*
+from ProcessInput import*
+from Classification_Annotation import*
+from Training import*
 import pickle
 
-
-def newData(audioDirectory, writeAddress, modelPath, generateDataReport=True, keepNPZFiles=True,
+def annotateNewData(audioDirectory, writeAddress, modelPath, fileList=None, generateDataReport=True, keepNPZFiles=True,
             numberOfMusicalExercises=5):
-    _ = fileWrite(audioDirectory, '', writeAddress)
 
-    newDataClassificationWrite(writeAddress, writeAddress, modelPath, generateDataReport, keepNPZFiles,
+    _ = writeFeatureData(audioDirectory, '', writeAddress, fileList)
+
+    classifyFeatureData(writeAddress, writeAddress, modelPath, generateDataReport, keepNPZFiles,
                                numberOfMusicalExercises)
 
 # FULL NEW DATA PROCESS ***************************************
 
-audioDirectory = "/Users/matthewarnold/Desktop/AutoSeg Local/AudioRepo/2014_CB_BariSax"
-npzDirectory = "/Users/matthewarnold/Desktop/AutoSeg Local/NPZ Repo/NewDataNPZ/2018/test"
-writeDirectory = "/Users/matthewarnold/Desktop/AutoSeg Local/TextOutput/2014_CB_BariSax"
-modelPath = '../Models/2017ABAI.sav'
+audioDirectory = "/Users/matthewarnold/Desktop/AutoSeg Local/AudioRepo/test"
+writeDirectory = "/Users/matthewarnold/Desktop/AutoSeg Local/TextOutput/test"
+modelPath = '../Models/2017ABAI.sav' #This is the model with best results from testing
 generateDataReport = True
 keepNPZFiles = False
 numberOfMusicalExercises = 5
 
-newData(audioDirectory, writeDirectory, modelPath, generateDataReport, keepNPZFiles, numberOfMusicalExercises)
-# newDataClassificationWrite(writeDirectory, writeDirectory, modelPath, generateDataReport, keepNPZFiles,
-#                                numberOfMusicalExercises)
+fileList = []
+
+annotateNewData(audioDirectory, writeDirectory, modelPath, fileList, generateDataReport, keepNPZFiles,
+                numberOfMusicalExercises)
+
+
+
+# TO TRAIN NEW MODEL *******************************************
+
+# # Directory for training data
+# trainingDirectory = "/Users/matthewarnold/Desktop/AutoSeg Local/AudioRepo/test"
+#
+# # Where model file should be written to and what filename
+# writeAddress = '../Models/'
+# modelFileName = "testModel"
+#
+# # If data is audio, True; if data is feature data npz, False
+# isAudio = True
+#
+# # If isAudio is true, textAddress for annotations of data is needed as well
+# textAddress = "/Users/matthewarnold/Desktop/AutoSeg Local/TextRepo/test"
+#
+# training(trainingDirectory, writeAddress, modelFileName, isAudio, textAddress)
